+++ lib/dartapi.dart
@@ -1,4 +1,5 @@
 export 'cli/create_project.dart';
 export 'cli/run_server.dart';
 export 'cli/generate_controllers.dart';
+export 'cli/generate_docs.dart';
 export 'constants/create_command_constants.dart';